refactor: 데이터베이스 모듈화, camelCase 사용

+++ controller/CartController.js
@@ -1,8 +1,7 @@
 const conn = require('../mariadb');
 const { StatusCodes } = require('http-status-codes');
 const jwt = require('jsonwebtoken');
-const dotenv = require('dotenv');
-dotenv.config();
+const ensureAuthorization = require('../auth'); // 인증모듈
 
 const addToCart = (req, res) => {
   const { book_id, quantity } = req.body;
@@ -44,11 +43,17 @@ const getCartItems = (req, res) => {
       message: `잘못된 토큰입니다.`,
     });
   } else {
-    let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price  
-              FROM cartItems LEFT JOIN books 
-              ON cartItems.book_id = books.id 
-              WHERE user_id = ? AND cartItems.id IN (?)`;
-    let values = [authorization.id, selected];
+    let sql = `SELECT cartItems.id, book_id, title, summary, quantity, price 
+    FROM cartItems LEFT JOIN books 
+    ON cartItems.book_id = books.id 
+    WHERE user_id = ?`;
+    let values = [authorization.id];
+
+    if (selected) {
+      // 주문서 작성 시 '선택한 장바구니 목록 조회'
+      sql += ` AND cartItems.id IN (?)`;
+      values.push(selected);
+    }
     conn.query(sql, values, (err, results) => {
       if (err) {
         console.log(err);
@@ -61,32 +66,26 @@ const getCartItems = (req, res) => {
 
 const removeCartItem = (req, res) => {
   const cartItemId = req.params.id;
+  let authorization = ensureAuthorization(req, res);
 
-  let sql = `DELETE FROM cartItems WHERE id = ?`;
-  conn.query(sql, cartItemId, (err, results) => {
-    if (err) {
-      console.log(err);
-      return res.status(StatusCodes.BAD_REQUEST).end();
-    }
-    return res.status(StatusCodes.OK).json(results);
-  });
-};
-
-function ensureAuthorization(req, res) {
-  try {
-    let receivedJwt = req.headers['authorization'];
-    console.log('receivedJwt : ', receivedJwt);
-
-    let decodedUserInfo = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
-    console.log(decodedUserInfo);
-
-    return decodedUserInfo;
-  } catch (err) {
-    console.log(err.name);
-    console.log(err.message);
-
-    return err;
+  if (authorization instanceof jwt.TokenExpiredError) {
+    return res.status(StatusCodes.UNAUTHORIZED).json({
+      message: `로그인 세션이 만료되었습니다. 다시 로그인 해주세요.`,
+    });
+  } else if (authorization instanceof jwt.JsonWebTokenError) {
+    return res.status(StatusCodes.BAD_REQUEST).json({
+      message: `잘못된 토큰입니다.`,
+    });
+  } else {
+    let sql = `DELETE FROM cartItems WHERE id = ?`;
+    conn.query(sql, cartItemId, (err, results) => {
+      if (err) {
+        console.log(err);
+        return res.status(StatusCodes.BAD_REQUEST).end();
+      }
+      return res.status(StatusCodes.OK).json(results);
+    });
   }
-}
+};
 
 module.exports = { addToCart, getCartItems, removeCartItem };
